+++ source/clientserver/error_log.cpp
@@ -48,7 +48,7 @@ void uda::client_server::print_errors(const std::vector<UdaError>& error_stack, 
     convert_non_printable2(access_date);
     trim_string(access_date);
 
-    for (int i = 0; i < request_block.size(); ++i) {
+    for (size_t i = 0; i < request_block.size(); ++i) {
         auto* request = &request_block[i];
         log(UDA_LOG_ERROR, __FILE__, __LINE__,
             "0 {} [{}] [{} {} {} {} {} {} {} {} {} {} {}]", client_block.uid, access_date,
