+++ source/clientserver/makeRequestBlock.cpp
@@ -700,7 +700,7 @@ int uda::client_server::make_request_block(const config::Config& config, Request
 {
     int rc = 0;
 
-    for (int i = 0; i < request_block->size(); ++i) {
+    for (size_t i = 0; i < request_block->size(); ++i) {
         auto* request = &(*request_block)[0];
         rc = make_request_data(config, request, pluginList);
         if (rc != 0) {
